Fix a few style nits with Diesel usage
These changes are mostly stylistic, but this code will now add `LIMIT 1`
to all the queries which are trying to get exactly one row. The changes
are:

- `table.filter(primary_key.eq(id))` => `table.find(id)`
- `.load()?.into_iter().nth(0)` => `.first().optional()`
  - Diesel treats "could not find a row" as an error by default,
    `.optional` just matches on `NotFound` and returns `None` instead
- Embed the migrations. If they're being run from the binary, there's no
  reason to require them on the file system to use it.

There were a few other structural issues that I didn't address, but
wanted to mention. All your `save_or_update` functions are super racy.
It looks like this uses PG, so this could use an
[upsert](http://docs.diesel.rs/diesel/pg/upsert/index.html) to avoid
that. At the very least, you should stick a `for_update` on those
selects (the ability to compose modifiers onto queries is why we usually
discourage writing functions that execute queries instead of returning
them). That would also require the call being in a transaction (which is
why we usually encourage passing a connection around explicitly)